chore: split code into emission_category_mapping service
